+++ pkg/storage/mysql/mysql.go
@@ -8,8 +8,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/openfga/openfga/pkg/encoder"
-
 	sq "github.com/Masterminds/squirrel"
 	"github.com/cenkalti/backoff/v4"
 	"github.com/go-sql-driver/mysql"
@@ -42,7 +40,6 @@ type Datastore struct {
 	dbInfo                 *sqlcommon.DBInfo
 	logger                 logger.Logger
 	dbStatsCollector       prometheus.Collector
-	tokenSerializer        encoder.ContinuationTokenSerializer
 	maxTuplesPerWriteField int
 	maxTypesPerModelField  int
 }
@@ -66,10 +63,6 @@ func New(uri string, cfg *sqlcommon.Config) (*Datastore, error) {
 		}
 
 		uri = dsnCfg.FormatDSN()
-	}
-
-	if cfg.TokenSerializer == nil {
-		cfg.TokenSerializer = sqlcommon.NewSQLContinuationTokenSerializer()
 	}
 
 	db, err := sql.Open("mysql", uri)
@@ -129,7 +122,6 @@ func NewWithDB(db *sql.DB, cfg *sqlcommon.Config) (*Datastore, error) {
 		db:                     db,
 		dbInfo:                 dbInfo,
 		logger:                 cfg.Logger,
-		tokenSerializer:        cfg.TokenSerializer,
 		dbStatsCollector:       collector,
 		maxTuplesPerWriteField: cfg.MaxTuplesPerWriteField,
 		maxTypesPerModelField:  cfg.MaxTypesPerModelField,
@@ -158,18 +150,13 @@ func (s *Datastore) Read(
 }
 
 // ReadPage see [storage.RelationshipTupleReader].ReadPage.
-func (s *Datastore) ReadPage(
-	ctx context.Context,
-	store string,
-	tupleKey *openfgav1.TupleKey,
-	options storage.ReadPageOptions,
-) ([]*openfgav1.Tuple, []byte, error) {
+func (s *Datastore) ReadPage(ctx context.Context, store string, tupleKey *openfgav1.TupleKey, options storage.ReadPageOptions) ([]*openfgav1.Tuple, string, error) {
 	ctx, span := startTrace(ctx, "ReadPage")
 	defer span.End()
 
 	iter, err := s.read(ctx, store, tupleKey, &options)
 	if err != nil {
-		return nil, nil, err
+		return nil, "", err
 	}
 	defer iter.Stop()
 
@@ -206,10 +193,7 @@ func (s *Datastore) read(ctx context.Context, store string, tupleKey *openfgav1.
 		sb = sb.Where(sq.Eq{"_user": tupleKey.GetUser()})
 	}
 	if options != nil && options.Pagination.From != "" {
-		token, _, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, err
-		}
+		token := options.Pagination.From
 		sb = sb.Where(sq.GtOrEq{"ulid": token})
 	}
 	if options != nil && options.Pagination.PageSize != 0 {
@@ -409,11 +393,7 @@ func (s *Datastore) ReadAuthorizationModel(ctx context.Context, store string, mo
 }
 
 // ReadAuthorizationModels see [storage.AuthorizationModelReadBackend].ReadAuthorizationModels.
-func (s *Datastore) ReadAuthorizationModels(
-	ctx context.Context,
-	store string,
-	options storage.ReadAuthorizationModelsOptions,
-) ([]*openfgav1.AuthorizationModel, []byte, error) {
+func (s *Datastore) ReadAuthorizationModels(ctx context.Context, store string, options storage.ReadAuthorizationModelsOptions) ([]*openfgav1.AuthorizationModel, string, error) {
 	ctx, span := startTrace(ctx, "ReadAuthorizationModels")
 	defer span.End()
 
@@ -425,10 +405,7 @@ func (s *Datastore) ReadAuthorizationModels(
 		OrderBy("authorization_model_id desc")
 
 	if options.Pagination.From != "" {
-		token, _, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, nil, err
-		}
+		token := options.Pagination.From
 		sb = sb.Where(sq.LtOrEq{"authorization_model_id": token})
 	}
 	if options.Pagination.PageSize > 0 {
@@ -437,7 +414,7 @@ func (s *Datastore) ReadAuthorizationModels(
 
 	rows, err := sb.QueryContext(ctx)
 	if err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 	defer rows.Close()
 
@@ -447,24 +424,21 @@ func (s *Datastore) ReadAuthorizationModels(
 	for rows.Next() {
 		err = rows.Scan(&modelID)
 		if err != nil {
-			return nil, nil, HandleSQLError(err)
+			return nil, "", HandleSQLError(err)
 		}
 
 		modelIDs = append(modelIDs, modelID)
 	}
 
 	if err := rows.Err(); err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 
-	var token []byte
+	var token string
 	numModelIDs := len(modelIDs)
 	if len(modelIDs) > options.Pagination.PageSize {
 		numModelIDs = options.Pagination.PageSize
-		token, err = s.tokenSerializer.Serialize(modelID, "")
-		if err != nil {
-			return nil, nil, err
-		}
+		token = modelID
 	}
 
 	// TODO: make this concurrent with a maximum of 5 goroutines. This may be helpful:
@@ -474,7 +448,7 @@ func (s *Datastore) ReadAuthorizationModels(
 	for i := 0; i < numModelIDs; i++ {
 		model, err := s.ReadAuthorizationModel(ctx, store, modelIDs[i])
 		if err != nil {
-			return nil, nil, err
+			return nil, "", err
 		}
 		models = append(models, model)
 	}
@@ -592,7 +566,7 @@ func (s *Datastore) GetStore(ctx context.Context, id string) (*openfgav1.Store, 
 }
 
 // ListStores provides a paginated list of all stores present in the storage.
-func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOptions) ([]*openfgav1.Store, []byte, error) {
+func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOptions) ([]*openfgav1.Store, string, error) {
 	ctx, span := startTrace(ctx, "ListStores")
 	defer span.End()
 
@@ -613,11 +587,7 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 		OrderBy("id")
 
 	if options.Pagination.From != "" {
-		token, _, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, nil, err
-		}
-		sb = sb.Where(sq.GtOrEq{"id": token})
+		sb = sb.Where(sq.GtOrEq{"id": options.Pagination.From})
 	}
 	if options.Pagination.PageSize > 0 {
 		sb = sb.Limit(uint64(options.Pagination.PageSize + 1)) // + 1 is used to determine whether to return a continuation token.
@@ -625,7 +595,7 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 
 	rows, err := sb.QueryContext(ctx)
 	if err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 	defer rows.Close()
 
@@ -636,7 +606,7 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 		var createdAt, updatedAt time.Time
 		err := rows.Scan(&id, &name, &createdAt, &updatedAt)
 		if err != nil {
-			return nil, nil, HandleSQLError(err)
+			return nil, "", HandleSQLError(err)
 		}
 
 		stores = append(stores, &openfgav1.Store{
@@ -648,18 +618,14 @@ func (s *Datastore) ListStores(ctx context.Context, options storage.ListStoresOp
 	}
 
 	if err := rows.Err(); err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 
 	if len(stores) > options.Pagination.PageSize {
-		contToken, err := s.tokenSerializer.Serialize(id, "")
-		if err != nil {
-			return nil, nil, err
-		}
-		return stores[:options.Pagination.PageSize], contToken, nil
+		return stores[:options.Pagination.PageSize], id, nil
 	}
 
-	return stores, nil, nil
+	return stores, "", nil
 }
 
 // DeleteStore removes a store from storage.
@@ -734,12 +700,7 @@ func (s *Datastore) ReadAssertions(ctx context.Context, store, modelID string) (
 }
 
 // ReadChanges see [storage.ChangelogBackend].ReadChanges.
-func (s *Datastore) ReadChanges(
-	ctx context.Context,
-	store string,
-	filter storage.ReadChangesFilter,
-	options storage.ReadChangesOptions,
-) ([]*openfgav1.TupleChange, []byte, error) {
+func (s *Datastore) ReadChanges(ctx context.Context, store string, filter storage.ReadChangesFilter, options storage.ReadChangesOptions) ([]*openfgav1.TupleChange, string, error) {
 	ctx, span := startTrace(ctx, "ReadChanges")
 	defer span.End()
 
@@ -767,15 +728,7 @@ func (s *Datastore) ReadChanges(
 		sb = sb.Where(sq.Eq{"object_type": objectTypeFilter})
 	}
 	if options.Pagination.From != "" {
-		token, objectType, err := s.tokenSerializer.Deserialize(options.Pagination.From)
-		if err != nil {
-			return nil, nil, err
-		}
-		if objectType != objectTypeFilter {
-			return nil, nil, storage.ErrMismatchObjectType
-		}
-
-		sb = sqlcommon.AddFromUlid(sb, token, options.SortDesc)
+		sb = sqlcommon.AddFromUlid(sb, options.Pagination.From, options.SortDesc)
 	}
 	if options.Pagination.PageSize > 0 {
 		sb = sb.Limit(uint64(options.Pagination.PageSize)) // + 1 is NOT used here as we always return a continuation token.
@@ -783,7 +736,7 @@ func (s *Datastore) ReadChanges(
 
 	rows, err := sb.QueryContext(ctx)
 	if err != nil {
-		return nil, nil, HandleSQLError(err)
+		return nil, "", HandleSQLError(err)
 	}
 	defer rows.Close()
 
@@ -808,14 +761,14 @@ func (s *Datastore) ReadChanges(
 			&insertedAt,
 		)
 		if err != nil {
-			return nil, nil, HandleSQLError(err)
+			return nil, "", HandleSQLError(err)
 		}
 
 		var conditionContextStruct structpb.Struct
 		if conditionName.String != "" {
 			if conditionContext != nil {
 				if err := proto.Unmarshal(conditionContext, &conditionContextStruct); err != nil {
-					return nil, nil, err
+					return nil, "", err
 				}
 			}
 		}
@@ -836,15 +789,10 @@ func (s *Datastore) ReadChanges(
 	}
 
 	if len(changes) == 0 {
-		return nil, nil, storage.ErrNotFound
+		return nil, "", storage.ErrNotFound
 	}
 
-	contToken, err := s.tokenSerializer.Serialize(ulid, objectTypeFilter)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return changes, contToken, nil
+	return changes, ulid, nil
 }
 
 // IsReady see [sqlcommon.IsReady].